refactor: streamline user balance retrieval by implementing From trait for UserBalancesReply #50
